Add Fraction class and update main program
A new class called Fraction is introduced which brings functionalities of creating and manipulating fractions. The Fraction class comes with methods for getting and setting numerator and denominator, displaying as string as well as calculating the decimal equivalent. Changes have also been made in the main program to test the newly implemented Fraction class.