Add scripture hiding functionality in the program
The main program functionality is changed from a simple "Hello World" message to a scripture hiding game. Users can now interact with the program to hide a custom number of words from a scripture verse. Also, the HideRandomWords function in the Scripture class has been improved to handle cases if the number of words to hide is greater than unhidden words.