Implement menu and improve activities experience
A new menu has been implemented providing the user with options to engage in different activities (breathing, reflecting and listing). Enhancements were added to the user experience for activities, including clearer instructions, countdowns, and the ability to set a specific session duration. Functionality has also been developed to randomly select prompts and questions in a smarter way for the reflecting activity.